Extracted data classes to separate module
