+++ actors.js
@@ -1,0 +1,8 @@
+class Stars {
+    constructor(name, link) {
+        this.name = name
+        this.link = link
+    }
+}
+
+module.exports = Stars
